Update README.md
This is a MAVEN project created using Spring MVC Framework.
This is a web based application, which is used to upload documents from the website to the Google Cloud.
While uploading the documents the user can also give the access rights and permission to the people or group, who can access this files.
Also, the user can provide specific folders in which he/she wanted to upload the file, and he/she can also create folders to upload files there.

To download certain file from the Google Cloud, user has to provide the file name which he wanted to download (The name can be partial, full name of the file is not required). The system will check the necessary permission for the file and if the user has the rights or belongs to a group which can access the file, then the file will be downloaded to their system.